# api/src/main/java/com/onesty/api/core/user/openapi (6).yaml
openapi: 3.0.0
info:
  title: Backend APIs
  description: APIs for the iOS app
  version: 1.0.0
servers:
  - url: https://api.example.com
paths:
  /users/{userId}:
    get:
      summary: Get user by ID
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /users:
    post:
      summary: Create a new user
      description: Register a new user in the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
...
components:
  schemas:
...
    UserBase:
      type: object
      properties:
        name:
          type: string
        gender:
          type: string
        age:
          type: integer
        height:
          type: integer
        weight:
          type: integer
        location:
          type: string
        profileName:
          type: string
        occupation:
          type: string
        languages:
          type: array
          items:
            type: string
        education:
          type: string
        children:
          type: string
        relationship:
          type: string
        sports:
          type: string
        alcohol:
          type: string
        smoking:
          type: string
        drugs:
          type: string
        activity:
          type: string
        aboutMe:
          type: string
        photos:
          type: array
          items:
            type: string
            format: uri
            maxItems: 5
    UserCreate:
      allOf:
        - $ref: '#/components/schemas/UserBase'
      required:
        - name
        - gender
        - age
        - height
        - weight
        - location
        - profileName
        - occupation
        - languages
        - education
        - children
        - relationship
...
    User:
      allOf:
        - $ref: '#/components/schemas/UserBase'
      properties:
        userId:
          type: string